+++ satella/coding/__init__.py
@@ -14,7 +14,7 @@ from .metaclasses import metaclass_maker, wrap_with, dont_wrap, wrap_property, D
     CopyDocsFrom
 from .misc import update_if_not_none, update_key_if_none, update_attr_if_none, queue_iterator, \
     update_key_if_not_none, source_to_function, update_key_if_true, \
-    get_arguments, call_with_locals
+    get_arguments, call_with_arguments
 from .recast_exceptions import rethrow_as, silence_excs, catch_exception, log_exceptions, \
     raises_exception
 from .overloading import overload, class_or_instancemethod
@@ -31,6 +31,6 @@ __all__ = [
     'rethrow_as', 'silence_excs', 'raises_exception',
     'static_var', 'metaclass_maker',
     'catch_exception', 'wraps', 'wrap_with', 'dont_wrap', 'wrap_property', 'log_exceptions',
-    'chain_functions', 'get_arguments', 'call_with_locals',
+    'chain_functions', 'get_arguments', 'call_with_arguments',
     'ListDeleter', 'DictDeleter', 'for_argument', 'attach_arguments', 'auto_adapt_to_methods',
 ]